+++ src/svs/inacademia_backend.py
@@ -50,6 +50,10 @@ class InAcademiaBackend(SAMLBackend):
         return result
 
     def authn_response(self, context, binding):
+        transaction_log(context.state, self.config.get("response_entry_order", 500),
+                        "inacademia_backend", "response", "entry", "success", '', '',
+                        'Recieved response from IdP')
+
         if not self.name in context.state:
             transaction_log(context.state, self.config.get("response_entry_order", 510),
                         "inacademia_backend", "response", "entry", "failed", '', '', 'Recieved response from IdP, but state lost', 'internal')
@@ -87,13 +91,12 @@ class InAcademiaBackend(SAMLBackend):
         if not internal_resp.user_id:
 
             transaction_log(state, self.config.get("response_exit_order", 620),
-                        "inacademia_backend", "response", "exit", "fail",resp_idp_entityid,'Failed to construct persistent user id from IdP response', 'idp')
+                            "inacademia_backend", "response", "exit", "fail", '', resp_idp_entityid,
+                            ErrorDescription.FAILED_TO_CONSTRUCT_PERSISTENT_USERID[LOG_MSG], 'idp')
 
-            raise SATOSAAuthenticationError(state, 'Failed to construct persistent user id from IdP response.')
-
-        # Happy Flow
-        transaction_log(state, self.config.get("response_entry_order", 500),
-                        "inacademia_backend", "response", "entry", "success", '', resp_idp_entityid, 'Recieved response from IdP')
+            auth_error = SATOSAAuthenticationError(state, "")
+            auth_error._message = ErrorDescription.FAILED_TO_CONSTRUCT_PERSISTENT_USERID[ERROR_DESC]
+            raise auth_error
 
         return internal_resp